iot-power-cc-monitor-py/requirements.txt
pyusb
pynvml~=11.5.0
psutil~=5.9.4
